functionality update to the bmp180 component class

+++ main/include/HandleBMP180.h
@@ -19,12 +19,13 @@ class HandleBMP180 : public HandleI2C //dont forget to put to private after test
     public:
     HandleBMP180();
     void BMP180_SetCalibrationData();
-    uint8_t BMP180_Read_16Bit(uint8_t DataAdress);
+    uint16_t BMP180_Read_16Bit(uint8_t DataAdress);
     uint8_t test();
 
     private:
     const uint8_t BMP180_I2C_ADRESS_Write = 0b11101110;
     const uint8_t BMP180_I2C_ADRESS_Read  = 0b11101111;
+
     short AC1;
     short AC2;
     short AC3;
